Move PulseMaker into HomeWorkspaceModel
Remove the reference to DynamoModel. Replace with a reference to
HomeWorkspaceModel. This needs to be improved.
